requirements.txt
streamlit>=1.24.0
pyrebase4>=4.7.1
firebase-admin>=6.2.0
requests>=2.31.0
gcloud>=0.18.3
python-jwt>=4.0.0
cryptography>=41.0.1
sseclient>=0.0.27
pycryptodome>=3.18.0

# Machine Learning dependencies
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
PyPDF2>=3.0.0
spacy>=3.6.0
textblob>=0.17.1